+++ src/qtel/ComDialog.h
@@ -186,7 +186,7 @@ class ComDialog : public ComDialogBase, public SigC::Object
     void updateStationData(const EchoLink::StationData *station);
     void createConnection(const EchoLink::StationData *station);
     void onStationListUpdated(void);
-    int micAudioRead(short *buf, int len);
+    int micAudioRead(float *buf, int len);
     bool openAudioDevice(Async::AudioIO::Mode mode);
 
   private slots:
